requirements.txt
streamlit
pypdf
python-docx
python-pptx
pandas
sentence-transformers
faiss-cpu
openai